Proposing changes to support key protection with Virtualization-based security to nf-ncrypt-ncryptcreatepersistedkey.md and ncryptimportkey
Hello, as part of planned general support of key protection with VBS (see aka.ms/keyguard), ENS KeyGuard team is wanting to update the documentation for NCryptCreatePersistedKey and NCryptImportKey. Specifically, these two functions should accept additional flags to request that VBS should be leveraged to protect the client key's private material: 

Flag 

NCRYPT_REQUIRE_VBS_FLAG 
Indicates a key must be created with VBS.   
Operation will fail if VBS is not available. 

NCRYPT_PREFER_VBS_FLAG 
Indicates a key should be created with VBS with a software fallback.  
Operation will attempt to fall back to software-only isolation if VBS is not available. 

Since this is a change coming in internally and we'd like to align with Marketing here, was wondering if we could get support from the page owner for suggestions on how to best format/go about this to get the change in January.

Thanks,
Emily